+++ CHANGES.rst
@@ -5,7 +5,7 @@ These features will be included in the next release:
 
 Added
 -----
-- Unit test for ``log_level`` option for `darkgraylib.config.load_config`.
+- Unit tests of configuration file options for `darkgraylib.config.load_config`.
 
 Fixed
 -----